renamings, refactorings, code review
