+++ qiskit_machine_learning/kernels/trainable_kernel_mixin.py
